Move compute_traj_pred_actual_actions, visualize, compute_mse to PolicyAlgo and remove validation in config_gen
