+++ robomimic/algo/algo.py
@@ -23,6 +23,7 @@ import robomimic.utils.torch_utils as TorchUtils
 import robomimic.utils.obs_utils as ObsUtils
 import robomimic.utils.action_utils as AcUtils
 
+from torch.utils.data import DataLoader
 
 # mapping from algo name to factory functions that map algo configs to algo class names
 REGISTERED_ALGO_FACTORY_FUNCS = OrderedDict()
@@ -417,6 +418,160 @@ class PolicyAlgo(Algo):
         """
         raise NotImplementedError
 
+    def compute_traj_pred_actual_actions(self, traj, return_images=False):
+        """
+        traj is an R2D2Dataset object representing one trajectory
+        This function is slow (>1s per trajectory) because there is no batching 
+        and instead loops through all timesteps one by one
+        TODO: documentation
+        """
+        if return_images:
+            image_keys = [item for item in traj.__getitem__(0)['obs'].keys() if "image" in item]
+            images = {key: [] for key in image_keys}
+        else:
+            images = None
+
+        dataloader = DataLoader(
+            dataset=traj,
+            sampler=None,
+            batch_size=1,
+            shuffle=False,
+            num_workers=1,
+            drop_last=True,
+        )
+
+        self.reset()
+        actual_actions = [] 
+        predicted_actions = [] 
+        
+        # loop through each timestep
+        for batch in iter(dataloader):
+            batch = self.process_batch_for_training(batch)
+
+            if return_images:
+                for image_key in image_keys:
+                    im = batch["obs"][image_key][0][-1]
+                    im = TensorUtils.to_numpy(im).astype(np.uint32)
+                    images[image_key].append(im)
+
+            batch = self.postprocess_batch_for_training(batch, obs_normalization_stats=None) # ignore obs_normalization for now
+
+            model_output = self.get_action(batch["obs"])
+            
+            actual_action = TensorUtils.to_numpy(
+                batch["actions"][0][0]
+            )
+            predicted_action = TensorUtils.to_numpy(
+                model_output[0]
+            )
+
+            actual_actions.append(actual_action)
+            predicted_actions.append(predicted_action)
+            
+        actual_actions = np.array(actual_actions)
+        predicted_actions = np.array(predicted_actions)
+        return actual_actions, predicted_actions, images
+    
+    def visualize(self, trainset, validset, savedir):
+        NUM_SAMPLES = 20
+
+        # set model into eval mode
+        self.set_eval()
+        random_state = np.random.RandomState(0)
+        train_indices = random_state.choice(
+            len(trainset.datasets),
+            min(len(trainset.datasets), NUM_SAMPLES)
+        ).astype(int)
+        training_sampled_data = [trainset.datasets[idx] for idx in train_indices]
+        
+        if validset is not None:
+            valid_indices = random_state.choice(
+                len(validset.datasets),
+                min(len(validset.datasets), NUM_SAMPLES)
+            ).astype(int)
+            validation_sampled_data = [validset.datasets[idx] for idx in valid_indices]
+        
+            inference_datasets_mapping = {"Train": training_sampled_data, "Valid": validation_sampled_data} 
+        else:
+            inference_datasets_mapping = {"Train": training_sampled_data}
+
+        action_names = self.get_action_names_for_vis(
+            action_keys=self.global_config.train.action_keys,
+            training_sample=training_sampled_data[0][0],
+        )
+
+        print("Saving model prediction plots to {}".format(savedir))
+
+        # loop through training and validation sets
+        for inference_key in inference_datasets_mapping:
+            # loop through each trajectory
+            traj_num = 1
+            for d in inference_datasets_mapping[inference_key]:
+                actual_actions, predicted_actions, images = self.compute_traj_pred_actual_actions(d, return_images=True)
+                save_path = os.path.join(savedir, "{}_traj_{}.png".format(inference_key, traj_num))
+                
+                self.make_model_prediction_plot(
+                    hdf5_path=d.hdf5_path,
+                    save_path=save_path,
+                    images=images,
+                    action_names=action_names,
+                    actual_actions=actual_actions,
+                    predicted_actions=predicted_actions,
+                )
+                traj_num += 1
+    
+    def compute_mse(self, trainset, validset):
+        NUM_SAMPLES = 32
+
+        # set model into eval mode
+        self.set_eval()
+        random_state = np.random.RandomState(0)
+        train_indices = random_state.choice(
+            len(trainset.datasets),
+            min(len(trainset.datasets), NUM_SAMPLES)
+        ).astype(int)        
+        training_sampled_data = [trainset.datasets[idx] for idx in train_indices]
+        
+        if validset is not None:
+            valid_indices = random_state.choice(
+                len(validset.datasets),
+                min(len(validset.datasets), NUM_SAMPLES)
+            ).astype(int)
+            validation_sampled_data = [validset.datasets[idx] for idx in valid_indices]
+        
+            inference_datasets_mapping = {"Train": training_sampled_data, "Valid": validation_sampled_data} 
+        else:
+            inference_datasets_mapping = {"Train": training_sampled_data}
+
+        
+        # loop through training and validation sets
+        for inference_key in inference_datasets_mapping:
+            actual_actions_all_traj = [] # (NxT, D)
+            predicted_actions_all_traj = [] # (NxT, D)
+
+            # loop through each trajectory
+            for d in inference_datasets_mapping[inference_key]:
+                actual_actions, predicted_actions, _ = self.compute_traj_pred_actual_actions(d)
+                actual_actions_all_traj.append(actual_actions)
+                predicted_actions_all_traj.append(predicted_actions)
+        
+        actual_actions_all_traj = np.concatenate(actual_actions_all_traj, axis=0)
+        predicted_actions_all_traj = np.concatenate(predicted_actions_all_traj, axis=0)        
+        accuracy_thresholds = np.logspace(-3,-5, num=3).tolist()
+        mse = torch.nn.functional.mse_loss(
+            torch.tensor(predicted_actions_all_traj), 
+            torch.tensor(actual_actions_all_traj), 
+            reduction='none'
+        ) # (NxT, D)
+        mse_log = {}
+        mse_log[f'{inference_key}/action_mse_error'] = mse.mean().item() # average MSE across all timesteps averaged across all action dimensions (D,)
+        
+        # compute percentage of timesteps that have MSE less than the accuracy thresholds
+        for accuracy_threshold in accuracy_thresholds:
+            mse_log[f'{inference_key}/action_accuracy@{accuracy_threshold}'] = (torch.less(mse,accuracy_threshold).float().mean().item())
+        
+        return mse_log
+        
 
 class ValueAlgo(Algo):
     """
